Add option to plot only certain labels in confusion matrices
Add an option to plot only certain specified labels in confusion
matrices to allow for cases where some "true" labels are not in the
"predicted" label set or vice versa.

This can be useful in cases where a classifier with certain labels is
applied to a dataset with a disjoint or partially disjoint set of
related labels.